move text-gen classes and functions to separate module
